Removed layers module
Custom layers should now be defined in model module and made specific to each project to keep the codebase lean
